+++ personnel/models.py
@@ -17,5 +17,5 @@ class Staff(models.Model):
     department_id = models.ForeignKey(Department, on_delete=models.CASCADE)
     staff_name = models.CharField(max_length=20, verbose_name='员工名字')
     staff_gender = models.CharField(max_length=2, verbose_name='员工性别')
-    staff_tel = models.IntegerField(max_length=11, verbose_name='员工电话')
+    staff_tel = models.PositiveBigIntegerField(max_length=11, verbose_name='员工电话')
     staff_address = models.CharField(max_length=50, verbose_name='员工家庭地址')
